09 01 changed to int container

+++ cpp09/ex01/RPN.hpp
@@ -8,7 +8,7 @@
 class RPN
 {
 	private:
-		std::stack<double>	_s;
+		std::stack<int>	_s;
 
 	public:
 		RPN();
